feat(api): add endpoint for getting all user data

+++ packages/api/src/modules/user/repositories/consumer-token.repository.ts
@@ -1,19 +1,24 @@
+import { ObjectId } from "mongodb";
 import db from "../../../shared/MongoConnection";
 import { ConsumerToken } from "../entities/ConsumerToken";
 
 class ConsumerTokenRepository {
     private readonly collection = db.collection<ConsumerToken>("consumer-token");
 
-    async findToken(userId) {
-        return (await this.collection.findOne({ userId }))?.token;
+    async findToken(userId: string | ObjectId) {
+        return (await this.collection.findOne({ userId: new ObjectId(userId) }))?.token;
+    }
+
+    async find(userId: string) {
+        return this.collection.find({ userId: new ObjectId(userId) }).toArray();
     }
 
     async create(entity: ConsumerToken) {
         return (await this.collection.insertOne(entity)).acknowledged;
     }
 
-    async deleteAllByUserId(userId) {
-        const result = await this.collection.deleteMany({ userId });
+    async deleteAllByUserId(userId: string | ObjectId) {
+        const result = await this.collection.deleteMany({ userId: new ObjectId(userId) });
         return result.acknowledged;
     }
 }
